+++ CHANGELOG.md
@@ -43,6 +43,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Fixed unable to piping the output of `Get-PnPRoleDefinition` to i.e. filter by RoleTypeKind.
 - Changed `Set-PnPListItemPermission` to support piping in a roledefinition for `-AddRole` and `-RemoveRole`
 - Fixed issue with 'Remove-PnPSiteDesign -Identity` not accepting a site design name, only a GUID.
+- Fixed issue with 'Remove-PnPSiteDesign -Identity` not accepting a site design name, only a GUID
+- Changed that `Get-PnPSiteScript -Identity` now also works with the site script name instead of just the site script Id
 
 ### Contributors